pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civilplan-mcp"
version = "1.0.0"
description = "CivilPlan MCP server for Korean civil and building project planning."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "22B Labs", email = "sinmb79@example.com" }
]
dependencies = [
  "fastmcp>=2.0.0",
  "pydantic>=2.0.0",
  "openpyxl>=3.1.0",
  "python-docx>=1.1.0",
  "svgwrite>=1.4.3",
  "ezdxf>=1.3.0",
  "aiosqlite>=0.20.0",
  "httpx>=0.27.0",
  "apscheduler>=3.10.0",
  "python-dateutil>=2.9.0",
]

[tool.setuptools.packages.find]
include = ["civilplan_mcp*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
